Move BinBuilder class to a separate file
